requirements.txt
pika
psycopg2-binary
git+https://github.com/ammaraskar/pyCraft@master#egg=minecraft
requests
future
cryptography
python-dotenv